% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duncan_fish_distribution_plots.R
\name{dunc_fish_distribution_plot}
\alias{dunc_fish_distribution_plot}
\title{Make plots to explore distribution of fish data}
\usage{
dunc_fish_distribution_plot(
  df = filter(fish_all, !is.na(Weight)),
  species = "KO",
  year = 2018,
  age = 2,
  variable = "Weight",
  ...,
  file_source = "DuncanGillnettData_2018 ages final.xlsx",
  plot = ggpubr::ggdensity
)
}
\arguments{
\item{df}{Dataframe cleaned up with nas removed from variable of interest}

\item{species}{Character value indicating fish species of interest. Defaults to KO}

\item{year}{Numeric value of year of the samples. Defaults to 2018}

\item{age}{Numeric value of age of the fish of interest}

\item{variable}{Character value indicating which column you wish to look at.  Defaults to weight}

\item{...}{Not used}

\item{file_source}{Name of source file where data came from. This is detailed in our database under the file_source column Defaults to the gillnet data only.}

\item{plot}{Type of plot to use. Fefaults to ggdensity. Other options include ggqqplot and others from the ggpubr package}
}
\description{
This function requires you to clean the input dataframe (df) to not include nas so
don't forget to filter them out!
}
