% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duncan_fish_data_tidy.R
\name{duncan_gillnet_tidy_2016}
\alias{duncan_gillnet_tidy_2016}
\title{Tidy 2016 gillnet datasets to prep for join with trawl data}
\usage{
duncan_gillnet_tidy_2016(df = gillnet_raw, sheet_name = "Duncan 2016 GN ")
}
\arguments{
\item{df}{List of dataframes each representing a sheet in DuncanGillnettData_2018 ages final.xlsx. This defaults to the gillnet_raw list object.}

\item{sheet_name}{String name of excel sheet in list loaded from original excel file. Defaults to "Duncan 2016 GN " which is the only sheet
this function will work on since the column names differ in the later years....}
}
\description{
Gillnet data loaded is from original excel workbook provided by FLNRO Large Lakes team (DuncanGillnettData_2018 ages final.xlsx).
The names of the columns have been repaired on import with .name_repair function from readxl \url{https://readxl.tidyverse.org/} package
The name of the sheet in the workbook is "Duncan 2016 GN "
}
