% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duncan_fish_data_tidy.R
\name{duncan_gillnet_tidy_2017}
\alias{duncan_gillnet_tidy_2017}
\title{Tidy 2017 gillnet dataset to prep for join with trawl data}
\usage{
duncan_gillnet_tidy_2017(df = gillnet_raw, sheet_name = "Duncan 2016 GN ")
}
\arguments{
\item{df}{List of dataframes each representing a sheet in DuncanGillnettData_2018 ages final.xlsx. This defaults to the gillnet_raw list object.}

\item{sheet_name}{String name of excel sheet in list loaded from original excel file. Defaults to "2017 GN".
Other option is "2018 GN"}
}
\description{
Gillnet data loaded is from original excel workbook provided by FLNRO Large Lakes team (DuncanGillnettData_2018 ages final.xlsx).
The names of the columns have been repaired on import with .name_repair function from \url{https://readxl.tidyverse.org/}  package
}
