% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duncan_fish_tables.R
\name{duncan_table_dunnTest}
\alias{duncan_table_dunnTest}
\title{Convert output of FSA::dunnTest to table for reporting.}
\usage{
duncan_table_dunnTest(x)
}
\arguments{
\item{x}{Captured output of FSA::dunnTest function}
}
\description{
Convert output of FSA::dunnTest to table for reporting and include unadjusted and adjusted p-values
as well as the method ued for adjustment.
}
