% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duncan_fish_tables.R
\name{duncan_table_kruskal}
\alias{duncan_table_kruskal}
\title{Convert output of stats::kruskal.test to table for reporting.}
\usage{
duncan_table_kruskal(x)
}
\arguments{
\item{x}{Captured output of stats::kruskal.test function}
}
\description{
Convert output of stats::kruskal.test to table for reporting.
}
\examples{
duncan_table_kruskal(kruskal.test(Ozone ~ Month, data = airquality))
}
