% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duncan_fish_data_tidy.R
\name{duncan_top2rows_to_colname}
\alias{duncan_top2rows_to_colname}
\title{Replace the column names of a dataframe by concatenating top two rows of each columns in an up-down direction.}
\usage{
duncan_top2rows_to_colname(df)
}
\arguments{
\item{df}{Dataframe object with details of the columns in the top two stacked rows as can be common in
excel style data presentation.  Note that the top two rows of the dataframe are removed after the column names are set.}
}
\description{
Used on the biomass data tidying "Duncan report_multi basin biomass.xlsx" file
but should work on any dataframe where the top two stacked rows contain details for a column
as can be common in excel style data presentation.
}
